+++ src/test/java/com/joyent/manta/client/MantaClientTest.java
@@ -52,7 +52,7 @@ public class MantaClientTest {
         ConfigContext config = new TestConfigContext(
                 mantaUrl, mantaUser, mantaKeyPath, mantaKeyId, mantaTimeout);
 
-        mantaClient = MantaClient.newInstance(config);
+        mantaClient = new MantaClient(config);
         testPathPrefix = String.format("/%s/stor/%s/", config.getMantaUser(), UUID.randomUUID());
         mantaClient.putDirectory(testPathPrefix, null);
     }